requirements copy.txt
accelerate==1.8.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio>=4.8.0,<5.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.2.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
brotlicffi==1.0.9.2
build==1.3.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.3.2
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
ddgs==9.5.4
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.17.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
duckdb==1.3.2
duckduckgo_search==5.3.1b1
durationpy==0.10
easychart==0.1.31
easytree==1.0.0
effdet==0.4.1
einops==0.8.1
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
Faker==19.13.0
fastapi>=0.115.0
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.5
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
func_timeout==4.3.5
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.2
google-genai>=1.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==0.49.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hjson==3.1.0
hpack==4.1.0
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.7
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==33.1.0
lagent==0.2.4
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-genai==2.1.12
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.11
lazy_loader==0.4
litellm==1.76.0
louis==1.3
lxml==6.0.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib>=3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.1
ml_dtypes==0.5.3
mmcv>=2.1.0
mmdet==3.3.0
mmengine==0.10.7
mmh3==5.2.0
mpi4py_mpich==3.1.5
mpmath==1.3.0
msgpack==1.1.1
msoffcrypto-tool==5.4.2
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.45.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1
nltk==3.9.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.575.51
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1
openai==1.102.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
peft==0.11.1
pexpect==4.9.0
phx-class-registry==4.1.0
pi_heif<1.1.0
pikepdf==9.10.2
pillow>=10.1.0,<11.0.0
pip==25.1
platformdirs==4.3.8
posthog==5.4.0
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow>=15.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
Pygments==2.19.2
Pylouis==0.0.1
pypandoc==1.15
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.14.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.4.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy>=1.11.4
seaborn==0.12.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==78.1.1
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
SQLAlchemy==1.4.54
sqlglot==25.34.1
sqlglotrs==0.3.0
squarify==0.4.4
stack-data==0.6.3
starlette>=0.37.0
streamlit==1.46.1
sympy==1.13.3
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timeout-decorator==0.5.0
timm==1.0.9
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.42.3
transformers-stream-generator==0.0.5
triton==2.3.1
typer==0.16.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unstructured==0.18.14
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.5
xtuner==0.1.23
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0

